chore(cleanup): move pure-math out of components/video-editor/videoPlayback/ into lib/zoomMath
Pure rename: the misnomer directory src/components/video-editor/videoPlayback/ contained 11 files of pure zoom/spring/focus math (zero React). Moved into a clearer home: src/lib/zoomMath/.

Files moved (9 source + 2 co-located unit tests):
  cameraFullscreenUtils.ts
  cameraFullscreenUtils.test.ts
  constants.ts
  cursorFollowUtils.ts
  focusUtils.ts
  mathUtils.ts
  motionSmoothing.ts
  zoomRegionUtils.ts
  zoomRegionUtils.test.ts (does not exist; the 2 tests are co-located above)
  zoomSpring.ts
  zoomSpring.test.ts
  zoomTransform.ts

Import sites updated (8 files, ~15 import specifiers):
  src/native/sceneDescription.test.ts
  src/lib/exporter/wgsl/evaluate.ts
  src/lib/exporter/frameRenderer.ts
  src/lib/cursor/pixiCursorRenderer.ts
  src/lib/cursor/cursorPathSmoothing.ts
  src/lib/ai-edition/timeline/zoom-preview.ts
  src/components/ai-edition/ZoomFocusOverlay.tsx
  src/components/ai-edition/PreviewCanvas.tsx

Internal relative imports of ../types rewritten to @/components/video-editor/types (5 sites: cameraFullscreenUtils, constants, cursorFollowUtils, focusUtils, zoomRegionUtils; plus the moved test). types.ts itself stays in components/video-editor/ — it is still a real shared types module used by src/lib/ai-edition/, and the parent directory src/components/video-editor/ stays in place; only the misnamed videoPlayback/ subfolder moved.

Doc references to the moved files (technical-documentation/architecture/cursor.md, crates/compositor/src/cursor.rs) updated to the new paths. Stale comments referencing files outside this move (VideoPlayback.tsx, videoPlayback/overlayUtils.ts, videoPlayback/videoEventHandlers.ts) are left alone — out of scope.

No behavior change. Verified with: tsc --noEmit (clean), vitest --run (1144/1144 passing), biome check (0 new warnings vs baseline).
