+++ src/lib/exporter/frameRenderer.ts
@@ -24,26 +24,6 @@ import {
 	isRotation3DIdentity,
 	lerpRotation3D,
 } from "@/components/video-editor/types";
-import { computeCameraFullscreenProgress } from "@/components/video-editor/videoPlayback/cameraFullscreenUtils";
-import {
-	AUTO_FOLLOW_PARAMS,
-	DEFAULT_FOCUS,
-} from "@/components/video-editor/videoPlayback/constants";
-import { advanceFollowFocus } from "@/components/video-editor/videoPlayback/cursorFollowUtils";
-import { clampFocusToScale } from "@/components/video-editor/videoPlayback/focusUtils";
-import { findDominantRegion } from "@/components/video-editor/videoPlayback/zoomRegionUtils";
-import {
-	createZoomSpringState,
-	resetZoomSpring,
-	stepZoomSpring,
-} from "@/components/video-editor/videoPlayback/zoomSpring";
-import {
-	applyZoomTransform,
-	computeFocusFromTransform,
-	computeZoomTransform,
-	createMotionBlurState,
-	type MotionBlurState,
-} from "@/components/video-editor/videoPlayback/zoomTransform";
 import {
 	computeCameraFullscreenRect,
 	computeCompositeLayout,
@@ -66,6 +46,19 @@ import {
 } from "@/lib/cursor/nativeCursor";
 import { BackgroundLoadError, classifyWallpaper, resolveImageWallpaperUrl } from "@/lib/wallpaper";
 import { drawCanvasClipPath } from "@/lib/webcamMaskShapes";
+import { computeCameraFullscreenProgress } from "@/lib/zoomMath/cameraFullscreenUtils";
+import { AUTO_FOLLOW_PARAMS, DEFAULT_FOCUS } from "@/lib/zoomMath/constants";
+import { advanceFollowFocus } from "@/lib/zoomMath/cursorFollowUtils";
+import { clampFocusToScale } from "@/lib/zoomMath/focusUtils";
+import { findDominantRegion } from "@/lib/zoomMath/zoomRegionUtils";
+import { createZoomSpringState, resetZoomSpring, stepZoomSpring } from "@/lib/zoomMath/zoomSpring";
+import {
+	applyZoomTransform,
+	computeFocusFromTransform,
+	computeZoomTransform,
+	createMotionBlurState,
+	type MotionBlurState,
+} from "@/lib/zoomMath/zoomTransform";
 import type { CursorRecordingData } from "@/native/contracts";
 import { renderAnnotations } from "./annotationRenderer";
 import {
