+++ src/lib/zoomMath/mathUtils.ts
