+++ src/lib/zoomMath/motionSmoothing.ts
