+++ src/lib/zoomMath/zoomRegionUtils.ts
@@ -1,6 +1,16 @@
+import type {
+	CursorTelemetryPoint,
+	Rotation3D,
+	ZoomFocus,
+	ZoomRegion,
+} from "@/components/video-editor/types";
+import {
+	DEFAULT_ROTATION_3D,
+	getRotation3D,
+	getZoomScale,
+	lerpRotation3D,
+} from "@/components/video-editor/types";
 import { clamp01 } from "@/utils/math";
-import type { CursorTelemetryPoint, Rotation3D, ZoomFocus, ZoomRegion } from "../types";
-import { DEFAULT_ROTATION_3D, getRotation3D, getZoomScale, lerpRotation3D } from "../types";
 import { TRANSITION_WINDOW_MS, ZOOM_IN_TRANSITION_WINDOW_MS } from "./constants";
 import { interpolateCursorAt } from "./cursorFollowUtils";
 import { clampFocusToScale } from "./focusUtils";
